new: nemo_debug
