+++ lib/comm/comm.cpp
@@ -1,22 +1,12 @@
 #include "comm.h"
-#include "setup.h"
+
 #include "wifi_credentials_private.h"
 
-// https://www.alejandrowurts.com/projects/esp32-wifi-udp/
-//  https://www.youtube.com/watch?v=U3vOyzJzGRU
-
-#include <WiFi.h>
-#include <WiFiUdp.h>
-
-#define N_BOXES 4
 #define M_CHARS 32
 
-WiFiUDP udp;         // Creation of wifi Udp instance
-WiFiUDP udp_monitor; // Creation of wifi Udp instance
+WiFiUDP udp; // Creation of wifi Udp instance
 
-char ip_address[] = UDP_ADDRESS;
-
-void wifi_setup() {
+void wifi_setup(const char *ip_address, int udp_port) {
     Serial.print("Trying to conecting to ");
 
     int n_networks = WiFi.scanNetworks();
@@ -60,46 +50,12 @@ void wifi_setup() {
         Serial.println("IP address: ");
         Serial.println(WiFi.localIP()); // 192.168.1.151
     }
-    udp.begin(UDP_PORT);
+    udp.begin(udp_port);
     Serial.println("Is this your local ip: " + (String)ip_address + "?");
 }
 
-void recieve_colour_code(char *colour_code, char local) {
-    char *message = NULL;
-
-    switch (local) {
-    case 'I':
-        send_data("IWP", UDP_ADDRESS);
-        break;
-    case 'O':
-        send_data("OWP", UDP_ADDRESS);
-        break;
-    case 'A':
-        send_data("MAP", UDP_ADDRESS);
-        break;
-    case 'B':
-        send_data("MBP", UDP_ADDRESS);
-        break;
-    }
-
-    Serial.println("Looking for next UDP package");
-    while (!message)
-        message = receive_data();
-    Serial.println("Valid package recieved from server...");
-
-    for (int i = 0; i < N_BOXES; i++) {
-        char aux = message[i];
-
-        if ((aux != 'R') & (aux != 'G') & (aux != 'B')) {
-            Serial.println("Invalid message for colour code");
-            colour_code[0] = '\0';
-            return;
-        }
-        colour_code[i] = aux;
-    }
-}
 char *receive_data() {
-    char *ret = (char *)calloc(M_CHARS, 1);
+    char *ret = (char *)calloc(32, 1);
     udp.parsePacket();
     if (udp.read(ret, M_CHARS) > 0) {
         Serial.print("Returning message from server: ");
@@ -123,30 +79,10 @@ char *recieve_data_impl() {
     return message;
 }
 
-void send_data(const char *message, const char *udpAddress) {
+void send_data(const char *message, const char *udpAddress, uint16_t udp_port) {
 
-    udp.beginPacket(udpAddress, UDP_PORT); // Initiate transmission of data
-
-    udp.printf(message);
-    udp.endPacket(); // Close communication
-}
-
-void debug_message(const char *message) {
-
-    if (!DEBUG_MODE)
-        return;
-    udp.beginPacket(UDP_ADDRESS, UDP_PORT); // Initiate transmission of data
+    udp.beginPacket(udpAddress, udp_port); // Initiate transmission of data
 
     udp.printf(message);
     udp.endPacket(); // Close communication
-
-    Serial.println(message);
-}
-
-void receive_colour_code_imp(char colour_code[4]) {
-
-    debug_message("Waiting for colour code...");
-    do {
-        recieve_colour_code(colour_code, 'I');
-    } while (colour_code[0] == '\0');
 }
