+++ src/nemo_motion/nemo_motion.h
@@ -7,3 +7,4 @@ void motors_setup();
 void walk(int millimeters, int direction, int turn);
 void rotate(int degrees, int direction, int speed);
 void correct_trajectory();
+void correct_trajectory_line();
